pyproject.toml
[tool.poetry]
name = "grpc-argument-validator"
version = "0.1.0"
description = ""
authors = [
    "Jos van de Wolfshaar <jos@messagebird.com>",
    "Jorn Engelbart <jorn@messagebird.com>"
]
packages = [
    { include = "grpc_argument_validator", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.6"
grpcio = "*"
protobuf = "*"

[tool.poetry.dev-dependencies]
grpcio-tools = "*"
pdoc3 = "^0.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
